requirements.txt
PySide6
PyYAML
numpy
gazu
Pillow